banco_dados/requirements.txt
astroid==2.4.2
autopep8==1.5.4
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector==2.2.9
pkg-resources==0.0.0
pycodestyle==2.6.0
pylint==2.6.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
wrapt==1.12.1
